requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
enum-compat==0.0.3
future==0.18.2
idna==2.10
Naked==0.1.31
parameterized==0.8.1
pycparser==2.20
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.25.1
serialization==1.2.3
shellescape==3.8.1
six==1.15.0
urllib3==1.26.3
zope.interface==5.2.0